pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "generative-models"
version = "0.1.0"
description = "Modular generative models playground"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "pytorch-lightning>=2.0",
    "hydra-core>=1.3",
    "hydra-submitit-launcher>=1.2",
    "omegaconf>=2.3",
    "wandb>=0.15",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "rich>=13.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
